Move our own files into `/app`
This makes it clearer which files are part of the tech-docs-template
and which are part of this project.
