+++ README.md
@@ -11,7 +11,8 @@ This is a static site generated with Middleman.
 This project uses [alphagov/tech-docs-template](https://github.com/alphagov/tech-docs-template).
 
 This means that some of the files (like the CSS, javascripts and layouts) are
-managed in the template and are not supposed to be modified here.
+managed in the template and are not supposed to be modified here. Any project-specific
+Ruby code needs to do into `/app`.
 
 You can pull down the latest version of the template by running:
