+++ app/content_schema.rb
