+++ app/dashboard.rb
