+++ app/publishing_api_docs.rb
