+++ app/supertypes.rb
