+++ lib/DONT_TOUCH.md
@@ -1,0 +1,4 @@
+The files in `/lib` are part of the [tech-docs-template][docs]. To keep us in
+sync with the template files here shouldn't be edited.
+
+[docs]: https://github.com/alphagov/tech-docs-template
