+++ spec/app/document_types_spec.rb
