+++ spec/app/github_spec.rb
