+++ spec/spec_helper.rb
@@ -1,6 +1,6 @@
 require 'webmock/rspec'
 
-require_relative './../lib/requires'
+require_relative './../app/requires'
 
 RSpec.configure do |config|
   config.before do
